refactor: 코드 리뷰 반영
extractDelimiter의 예외 처리 변경
- null을 반환하지 않고 DelimiterNotFoundException을 발생시킵니다
- 예외발생에 따라 함수명을 변경합니다

inner class를 별도 분리

사용하지 않는 함수 제거
