+++ src/main/java/stringcalculator/DelimiterNotFoundException.java
@@ -1,0 +1,5 @@
+package stringcalculator;
+
+public class DelimiterNotFoundException extends RuntimeException {
+
+}
